backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic-settings==2.3.1
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
aiosqlite==0.20.0
anthropic>=0.74.1
pgvector==0.4.1
tiktoken==0.12.0
